Moving M3x3Float defintion to data folder and static_matrix.h file. Improves include hierarchies, avoiding to include algebraic operations in places where they are not needed
