+++ include/fused_kernel/algorithms/image_processing/color_conversion.h
@@ -21,6 +21,7 @@
 #include <fused_kernel/algorithms/image_processing/saturate.h>
 #include <fused_kernel/algorithms/image_processing/raw_image.h>
 #include <fused_kernel/algorithms/image_processing/itu_color.h>
+#include <fused_kernel/algorithms/basic_ops/algebraic.h>
 #include <fused_kernel/algorithms/basic_ops/vector_ops.h>
 
 namespace fk {
